deleted old text area rule and renamed the input rule
