+++ README.md
@@ -162,7 +162,7 @@ Any use of third-party trademarks or logos are subject to those third-party's po
 | [dropdown-needs-labelling](docs/rules/dropdown-needs-labelling.md)                                                     | Accessibility: Dropdown menu must have an id and it needs to be linked via htmlFor of a Label                                                     | ✅  |    |    |
 | [image-button-missing-aria](docs/rules/image-button-missing-aria.md)                                                   | Accessibility: Image buttons must have accessible labelling: title, aria-label, aria-labelledby, aria-describedby                                 | ✅  |    |    |
 | [image-link-missing-aria](docs/rules/image-link-missing-aria.md)                                                       | Accessibility: Image links must have an accessible name                                                                                           | ✅  |    | 🔧 |
-| [input-missing-label](docs/rules/input-missing-label.md)                                                               | Accessibility: Input fields must have accessible labelling: aria-label, aria-labelledby or an associated label                                    | ✅  |    |    |
+| [input-components-require-accessible-name](docs/rules/input-components-require-accessible-name.md)                     | Accessibility: Input fields must have accessible labelling: aria-label, aria-labelledby or an associated label                                    | ✅  |    |    |
 | [menu-item-needs-labelling](docs/rules/menu-item-needs-labelling.md)                                                   | Accessibility: MenuItem without label must have an accessible and visual label: aria-labelledby                                                   | ✅  |    |    |
 | [no-empty-buttons](docs/rules/no-empty-buttons.md)                                                                     | Accessibility: Button, ToggleButton, SplitButton, MenuButton, CompoundButton must either text content or icon or child component                  | ✅  |    |    |
 | [no-empty-components](docs/rules/no-empty-components.md)                                                               | FluentUI components should not be empty                                                                                                           | ✅  |    |    |
@@ -173,7 +173,6 @@ Any use of third-party trademarks or logos are subject to those third-party's po
 | [spin-button-unrecommended-labelling](docs/rules/spin-button-unrecommended-labelling.md)                               | Accessibility: Unrecommended accessibility labelling - SpinButton                                                                                 | ✅  |    |    |
 | [spinner-needs-labelling](docs/rules/spinner-needs-labelling.md)                                                       | Accessibility: Spinner must have either aria-label or label, aria-live and aria-busy attributes                                                   | ✅  |    |    |
 | [switch-needs-labelling](docs/rules/switch-needs-labelling.md)                                                         | Accessibility: Switch must have an accessible label                                                                                               | ✅  |    |    |
-| [text-area-missing-label](docs/rules/text-area-missing-label.md)                                                       | Accessibility: Textarea must have an accessible name                                                                                              | ✅  |    |    |
 | [toolbar-missing-aria](docs/rules/toolbar-missing-aria.md)                                                             | Accessibility: Toolbars need accessible labelling: aria-label or aria-labelledby                                                                  | ✅  |    |    |
 | [tooltip-not-recommended](docs/rules/tooltip-not-recommended.md)                                                       | Accessibility: Prefer text content or aria over a tooltip for these components MenuItem, SpinButton                                               | ✅  |    |    |
