+++ lib/applicableComponents/inputBasedComponents.js
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-const applicableComponents = ["Input", "Slider", "DatePicker", "TextArea", "TextField", "TimePicker", "SearchBox"];
+const applicableComponents = ["Input", "Slider", "DatePicker", "Textarea", "TextField", "TimePicker", "SearchBox"];
 
 module.exports = {
     applicableComponents
