+++ lib/index.js
@@ -17,10 +17,9 @@ module.exports = {
         "checkbox-needs-labelling": require("./rules/checkbox-needs-labelling"),
         "image-button-missing-aria": require("./rules/image-button-missing-aria"),
         "image-link-missing-aria": require("./rules/image-link-missing-aria"),
-        "input-missing-label": require("./rules/input-missing-label"),
+        "input-components-require-accessible-name": require("./rules/input-components-require-accessible-name"),
         "menu-item-needs-labelling": require("./rules/menu-item-needs-labelling"),
         "switch-needs-labelling": require("./rules/switch-needs-labelling"),
-        "text-area-missing-label": require("./rules/text-area-missing-label"),
         "toolbar-missing-aria": require("./rules/toolbar-missing-aria"),
         "combobox-needs-labelling": require("./rules/combobox-needs-labelling"),
         "no-empty-components": require("./rules/no-empty-components"),
@@ -48,10 +47,9 @@ module.exports = {
             rules: {
                 "@microsoft/fluentui-jsx-a11y/checkbox-needs-labelling": "error",
                 "@microsoft/fluentui-jsx-a11y/image-link-missing-aria": "error",
-                "@microsoft/fluentui-jsx-a11y/input-missing-label": "error",
+                "@microsoft/fluentui-jsx-a11y/input-components-require-accessible-name": "error",
                 "@microsoft/fluentui-jsx-a11y/menu-item-needs-labelling": "error",
                 "@microsoft/fluentui-jsx-a11y/switch-needs-labelling": "error",
-                "@microsoft/fluentui-jsx-a11y/text-area-missing-label": "error",
                 "@microsoft/fluentui-jsx-a11y/image-button-missing-aria": "error",
                 "@microsoft/fluentui-jsx-a11y/toolbar-missing-aria": "error",
                 "@microsoft/fluentui-jsx-a11y/combobox-needs-labelling": "error",
