--- lib/rules/text-area-missing-label.js
@@ -1,57 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-"use strict";
-
-var elementType = require("jsx-ast-utils").elementType;
-const { hasFieldParent } = require("../util/hasFieldParent");
-const { isInsideLabelTag, hasAssociatedLabelViaHtmlFor, hasAssociatedLabelViaAriaLabelledBy } = require("../util/labelUtils");
-
-//------------------------------------------------------------------------------
-// Rule Definition
-//------------------------------------------------------------------------------
-
-module.exports = {
-    meta: {
-        // possible error messages for the rule
-        messages: {
-            missingLabel: "Textarea must have a label associated with it"
-        },
-        // "problem" means the rule is identifying code that either will cause an error or may cause a confusing behavior: https://eslint.org/docs/latest/developer-guide/working-with-rules
-        type: "problem",
-        // docs for the rule
-        docs: {
-            description: "Accessibility: Textarea must have an accessible name",
-            recommended: true,
-            url: "https://www.w3.org/TR/html-aria/" // URL to the documentation page for this rule
-        },
-        schema: []
-    },
-    // create (function) returns an object with methods that ESLint calls to “visit” nodes while traversing the abstract syntax tree
-    create(context) {
-        return {
-            // visitor functions for different types of nodes
-            JSXOpeningElement(node) {
-                // if it is not a Textarea, return
-                if (elementType(node) !== "Textarea") {
-                    return;
-                }
-
-                // if the Textarea is inside a Label tag, return
-                if (
-                    hasFieldParent(context) ||
-                    isInsideLabelTag(context) ||
-                    hasAssociatedLabelViaHtmlFor(node, context) ||
-                    hasAssociatedLabelViaAriaLabelledBy(node, context)
-                ) {
-                    return;
-                }
-
-                context.report({
-                    node,
-                    messageId: `missingLabel`
-                });
-            }
-        };
-    }
-};
